Consolidate Galata tests into single notebook and test file
Merge 5 separate test notebooks into test_all.ipynb and 3 test files
into megane.test.ts to reduce JupyterLab startup/kernel overhead.
Use test.describe.serial with shared page to run all cells once.
Add trusted:true metadata to eliminate "not trusted" warnings.
